src/requirements.txt
Django==2.1.15
djangorestframework==3.8.2
flake8==3.6.0
django-cors-headers==3.4.0